[RF] Allow for const template types in RooTemplateProxy.
Since the type-safe and const-correct template proxy derives from an
abstract proxy that only stores RooAbsArg*, we need to introduce a
const_cast when calling a base class constructor.
The const correctness is only enforced in the derived class, but this is
the one that's providing operator->, where const correctness matters.

When the proxy is used polymorphically via a RooAbsProxy*, constness
indeed vanishes, but that's correct, because operations on the
computation graph might have to do that anyway. This is similar to
directly accessing the original object that's stored in a
unique_ptr<const Type>.