[metacling] Access of a ctor is access of the base:
A ctor pulled in through using decls is public / protected / private
if the base nominated by the ctor is public / protected / private.
The location of the using declaration within the derived class
has no influence on the access. Of course a constructor that is
private within the base class is still private in the derived
class, whatever the inheritance is.