[RF] Relax checking of batch computation results in RooNLLVar.
If ROOFIT_CHECK_CACHED_VALUES is #defined, there was a hard assertion
failure when running a safety check that batch computations yield the
same result as single-value computations with a tolerance of 1.E-13

This has been relaxed. Now, only a message will be printed. A hard
failure only happens when values deviate more than 1.E-9