When using externalGradient use still Numerical gradient as input to SeedGenerator
With external gradient one does not provide often correct step sizes and therefore the initial state computed with SeedGenerator could be  really off and many more iterations will be afterwards required.
It is much more convenient to use once the numerical gradient to compute correct step sizes and initial seed. In particular by using the numerical gradient one has also for free the second derivatives (the diagonal part of the Hessian)

A long term solution would be to provide also as input the second derivatives.