[RF] Consider rangeName for integration in RooParamHistFunc
The logic for summing over histogram bins in different ranges used in
RooHistPdf is also implemented in RooParamHistFunc. This means the
range is now considered when computing integrals of RooParamHistFunc.

RooParamHistFunc allows you to scale the counts in each bin with a
parameter. The interface of RooDataHist::sum was extended with a
function parameter to inject the logic of scaling the bin weight
depending on the bin index.

This commit partly fixes issue #7182. We still need to implement the
range feature in RooHistFunc.